src/backend/requirements.txt
mysql-connector-python
flask
PyJWT
argon2